wal: commit queue and commit fiber
Do not commit a journal entry after write but put it into a
commit queue. This a preparation for synchronous replication
implementation. At this time commit queue is processed by a commit
fiber which commits and schedules each entry whereas in the
near future the fiber will check majority before commit.

Part of #980