Libev update to version 4.32
There was a bug in libev that caused some stress tests to fail
with an error indicating lack of fds on mac. The flag which was
supposed to fix the issue (DARWIN_UNLIMITED_SELECT) was defined
too late in earlier versions of libev.

More precisely, it was defined after including time.h which in
its turn had this line:
include <sys/_select.h>    /* select() prototype */

And <sys/_select.h> had this:
if defined(_DARWIN_C_SOURCE) || defined(_DARWIN_UNLIMITED_SELECT)
        __DARWIN_EXTSN_C(select)

So unlimited select flag did not do the trick as it was supposed to.
It was fixed in 4.25 along with other changes.

Closes #3867
Closes #4673